requirements.txt
APScheduler==3.8.1
python-dotenv==0.19.1
requests==2.26.0
beautifulsoup4==4.10.0
fake-useragent==0.1.11
python-telegram-bot==19.8
